requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
beautifulsoup4==4.14.3
ccxt==4.5.36
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coincurve==21.0.0
colorama==0.4.6
cryptography==46.0.4
frozenlist==1.8.0
idna==3.11
joblib==1.5.3
llvmlite==0.44.0
multidict==6.7.1
nltk==3.9.2
numba==0.61.2
numpy==2.2.6
pandas==3.0.0
pandas-ta==0.4.71b0
propcache==0.4.1
pycares==5.0.1
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
regex==2026.1.15
requests==2.32.5
schedule==1.2.2
setuptools==80.10.2
six==1.17.0
soupsieve==2.8.3
textblob==0.19.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
yarl==1.22.0
